refactor(mentors): split character_arc.py into smaller modules to fix file size limit
- Reduce character_arc.py from 2538 to 1006 lines (within 1500 limit)
- Extract relationship analysis to character_arc_relationships.py (608 lines)
- Extract arc types and constants to character_arc_types.py (181 lines)
- Extract transformation markers to character_arc_markers.py (236 lines)
- Extract conflict analysis to character_arc_conflict_analysis.py (243 lines)
- Extract arc detection utilities to character_arc_detection.py (257 lines)
- Extract TypedDict definitions to character_arc_analysis_types.py (34 lines)
- Update tests to use new module structure
- All files now comply with MCP tool size limits
- All tests passing (23/23)
